Move TripOnServiceDate to timetable package
