+++ src/main/java/org/opentripplanner/netex/mapping/TripPatternMapperResult.java
@@ -7,10 +7,10 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import org.opentripplanner.model.StopTime;
-import org.opentripplanner.model.TripOnServiceDate;
 import org.opentripplanner.transit.model.network.StopPattern;
 import org.opentripplanner.transit.model.network.TripPattern;
 import org.opentripplanner.transit.model.timetable.Trip;
+import org.opentripplanner.transit.model.timetable.TripOnServiceDate;
 
 /**
  * This mapper returnes two collections, so we need to use a simple wraper to be able to return the
