+++ src/main/java/org/opentripplanner/routing/DatedServiceJourneyHelper.java
@@ -3,8 +3,8 @@ package org.opentripplanner.routing;
 import java.time.LocalDate;
 import org.opentripplanner.model.TimetableSnapshot;
 import org.opentripplanner.model.TripIdAndServiceDate;
-import org.opentripplanner.model.TripOnServiceDate;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
+import org.opentripplanner.transit.model.timetable.TripOnServiceDate;
 import org.opentripplanner.transit.service.TransitService;
 
 public class DatedServiceJourneyHelper {
