+++ src/main/java/org/opentripplanner/transit/model/timetable/TripOnServiceDate.java
@@ -1,11 +1,9 @@
-package org.opentripplanner.model;
+package org.opentripplanner.transit.model.timetable;
 
 import java.time.LocalDate;
 import java.util.List;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 import org.opentripplanner.transit.model.framework.TransitEntity;
-import org.opentripplanner.transit.model.timetable.Trip;
-import org.opentripplanner.transit.model.timetable.TripAlteration;
 
 /**
  * Class for holding data about a certain trip on a certain day. Essentially a DatedServiceJourney.
